chore: move sudo child process code to Linux dir for conditional coverage
